// math-common/build/notationTypes.d.ts
export type Optional<T, K extends keyof T> = Omit<T, K> & Partial<Pick<T, K>>;
export interface BaseNotation {
    id: number;
    uuid: string;
    user: UserAttributes;
    notationType: NotationType;
    boardType: BoardType;
    createdAt: Date;
}
// ...
export interface UserAttributes {
    id?: number;
    uuid: string;
    firstName: string;
    lastName: string;
    email: string;
    password: string;
    access_token: string;
    imageUrl: string;
    userType: UesrType;
    authorized: boolean;
    lastHeartbeatTime: Date;
}
export interface UserCreateAttributes extends Optional<UserAttributes, "id"> {
}
export interface AnswerAttributes extends BaseNotation {
    name: string;
    questionId: number;
    question: QuestionAttributes;
}
export interface AnswerCreateAttributes extends Optional<AnswerAttributes, "id"> {
}
// ...
export interface QuestionAttributes extends BaseNotation {
    lessonId: number;
    lesson: LessonAttributes;
    name: string;
}
// ...
export interface LessonAttributes extends BaseNotation {
    name: string;
}
// ...
export interface StudentLessonAttributes {
    id?: number;
    user: UserAttributes;
    userId: number;
    lessonId: number;
    lesson: LessonAttributes;
}
export interface StudentLessonCreateAttributes extends Optional<StudentLessonAttributes, "id"> {
}